[FEATURE]: CorrectInDIaryDetail 제작->본문 내용
